pyproject.toml
[tool.poetry]
name = "github-org-scripts"
version = "0.1.0"
description = "Misc scripts"
authors = ["Hal Wine <hwine@mozilla.com>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.8"
jsonstreams = "=0.4.1"
PyYaml = "=5.4"
tinydb = "=3.2.1"
github_selenium = { git = "https://github.com/mozilla/github_selenium.git" }
arrow = "^0.12.1"
#"github3.py" = { git = "https://github.com/hwine/github3.py.git", branch = "invitations" }
jupyter = "^1.0.0"
argcomplete = "^1.12.3"
"github3.py" = "<2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
jupyterlab = "^3.0.14"
